multiple(backend)
fix(handlers:answer): tuned handlers to tests

feat(repositories:answer): made totalize and post answer return sql.ErrNoRows

feat(testRunner): using more memory efficient values and not ptrs to values(they are not modified anyway)

feat(middleware): added Running middleware!

fix(quizParsers): removed excesive logging

feat(migrations): made score REAL

fix(server): multiple pathes simplification/middlewarification/fixing

+++ application/internal/database/migrations/postgres/20260530134723_created.up.sql
@@ -54,7 +54,7 @@ CREATE TABLE IF NOT EXISTS "users_groups_tests" (
     "test_uuid"  UUID REFERENCES tests(uuid) ON DELETE CASCADE,
     "group_uuid" UUID REFERENCES groups(uuid) ON DELETE CASCADE,
     "user_uuid"  UUID REFERENCES users(uuid) ON DELETE CASCADE,
-    "score"      SMALLINT NOT NULL,
+    "score"      FLOAT NOT NULL,
     
     PRIMARY KEY("test_uuid","group_uuid","user_uuid")
 );
